Valid URLs generated against default base URL
The previous commit just updated the impl and the unit tests so the
level of testing was similar to before. But obviously even though the
tests were passing before the URL generation was wrong.

To up our testing lets do a smoke test to verify that the URLs generated
for Selenium actually resolve to a valid URL. This of course doesn't
say anything about what data is at that URL but at least the resource
is valid.

Thought about extending this to the other downloads (web drivers) but
figured I would keep the scope to just what was failing. If someone
wants to extend this it should be fairly straightforward to do
something similar.