Refactor writing of downloaded files
The primary goal is to have a single function that handles the writing
so we have a single place where we can save the etag from the HTTP
response.

This also DRYs up some existing code including fixing a small issue
with `installSingleFile`. The other `install` functions all correctly
install the `error` event handler on the writableStream. The
`installSingleFile` installs it on the download stream (which already
has one) and leaves the writableStream without