Partially revert 2da8f60 (checking only top-level folder)
When a file within a folder is updated the folder does not change it's
timestamp. Only if a new file/folder is added/removed will the
timestamp change.

The commit message on 2da8f60 doesn't make clear if the developer
thought the top-level folder would update the timestamp but that is
my assumption. It's also not clear what failure was happening with the
test but my assumption is what it was the etag issue I'm work to
correct.

This commit updates the test so it does the folder walk it was doing
previously and verifies that ALL files have the same timestamp. Since
the previous code to walk the tree was made a while back I did make
some style changes to match with current project conventions.