Use etag for conditional HTTP request
The previous commit updated the system so the etag is saved next to
the file downloaded. This actually uses that info to make a proper
conditional HTTP request.

With this in place we can actually remove the special exception for
Edge as well. It wasn't that conditional HTTP requests were not
working but just that the MD5 hash was not the right tag to send for
that server.