Fix Conditional HTTP Requests (#667)
* Partially revert 2da8f60 (checking only top-level folder)

When a file within a folder is updated the folder does not change it's
timestamp. Only if a new file/folder is added/removed will the
timestamp change.

The commit message on 2da8f60 doesn't make clear if the developer
thought the top-level folder would update the timestamp but that is
my assumption. It's also not clear what failure was happening with the
test but my assumption is what it was the etag issue I'm work to
correct.

This commit updates the test so it does the folder walk it was doing
previously and verifies that ALL files have the same timestamp. Since
the previous code to walk the tree was made a while back I did make
some style changes to match with current project conventions.

* Refactor writing of downloaded files

The primary goal is to have a single function that handles the writing
so we have a single place where we can save the etag from the HTTP
response.

This also DRYs up some existing code including fixing a small issue
with `installSingleFile`. The other `install` functions all correctly
install the `error` event handler on the writableStream. The
`installSingleFile` installs it on the download stream (which already
has one) and leaves the writableStream without

* Record etag when saving downloaded stream

This will allow for a conditional HTTP request later.

Using the FS promise API is simpler than the callback API so I used that
for writeFile but to avoid `fs` confusion just extracted the relevant
methods from each module.

* Use etag for conditional HTTP request

The previous commit updated the system so the etag is saved next to
the file downloaded. This actually uses that info to make a proper
conditional HTTP request.

With this in place we can actually remove the special exception for
Edge as well. It wasn't that conditional HTTP requests were not
working but just that the MD5 hash was not the right tag to send for
that server.

* Last minute change made lint complain. Fix

* Make compat with v12 of NodeJS

In NodeJS v12 it is:

  require('fs').promises

instead of

  require('fs/promises')