Update test suite to detect when Chrome didn't load
The commit with a problem caused no drivers to load but the tests
did not detect this as they were looking to see what args were
specified not what was the result of those args with regard to what
drivers loaded.

This commit updates the test suite to better detect the problem. It
captures the stdout and checks it for Chrome instead of just checking
the args.

With this in place I now have a failing test with my problematic code
from before.