Make selenium use only the drivers specified
This removes the previous attempt to do this, `--detect-drivers`. That
option would be better named as "--load-drivers" as it won't load any
drivers regardless if specified the path via the CLI.

According to this discussion:

https://github.com/SeleniumHQ/selenium/issues/8928

The `-I` driver is what is needed to specify what drivers should be
loaded. Therefore for each driver we are providing an explicit path we
are also indicating that driver should be loaded. Any without a -I
will not get loaded.

I updated the failing test so it does explicitly load the firefox driver
(to avoid the jar throwing an error about no drivers being loaded) and
then have the test verify that Chrome is not in the output indicating
that other drivers such as Chrome did not load.

The possible values for `-I` doesn't seem well documented (that I could
find). The code seems to compare the CLI value to the "display name"
for each driver info class:

https://github.com/SeleniumHQ/selenium/blob/9546dba7a5aefe45ff8fe2bbc1039819dde10672/java/src/org/openqa/selenium/grid/node/config/NodeOptions.java#L460

Therefore to get a complete list of the right values I just looked at
all the implementations of the "getDisplayName" method.

I have personally manually tested chrome, firefox and chromium edge.
I don't have a win32 system to test IE and the legacy version of
Edge but based on the display name it *should* work just fine.